requirements.txt
Flask==0.12.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
psycopg2==2.7.1
Flask-Testing==0.6.2
mysqlclient
flask-cors==3.0.2
pandas
numpy
Flask-WTF==0.14.2
WTForms==2.1
flask-login
